@@ -2,36 +2,37 @@
     "name": "example-spa",
     "version": "1.0.0",
     "description": "An example SPA that uses Curity recommendations for web security",
+    "author": "Curity AB",
+    "license": "Apache-2.0",
     "main": "index.ts",
+    "type": "module",
     "engines": {
-        "node": ">=18"
+        "node": ">=20"
     },
     "scripts": {
         "start": "npm run clean && npm run copy && npm run serve",
-        "clean": "rimraf dist",
+        "clean": "rm -rf dist",
         "copy": "mkdir dist && cp index.html app.css config.json dist/",
-        "serve": "webpack serve --config webpack/webpack.dev.js",
-        "build": "npm run clean && npm run copy && webpack --config webpack/webpack.prod.js",
+        "serve": "NODE_OPTIONS='--import tsx' webpack serve --config webpack/webpack.dev.ts",
+        "build": "npm run clean && npm run copy && NODE_OPTIONS='--import tsx' webpack --config webpack/webpack.prod.ts",
         "uitests": "cypress run --headed --no-exit --browser chrome"
     },
-    "author": "Curity AB",
-    "license": "Apache-2.0",
     "dependencies": {
-        "axios": "^1.4.0",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0"
+        "axios": "^1.7.2",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1"
     },
     "devDependencies": {
-        "@types/react": "^18.0.26",
-        "@types/react-dom": "^18.0.9",
-        "cypress": "^13.6.2",
+        "@types/react": "^18.3.2",
+        "@types/react-dom": "^18.3.0",
+        "cypress": "^13.10.0",
         "mochawesome": "^7.1.3",
-        "rimraf": "^3.0.2",
         "ts-loader": "^9.4.2",
-        "typescript": "^5.1.5",
-        "webpack": "^5.88.1",
+        "tsx": "^4.11.0",
+        "typescript": "^5.4.5",
+        "webpack": "^5.91.0",
         "webpack-cli": "^5.1.4",
-        "webpack-dev-server": "^4.15.1",
+        "webpack-dev-server": "^4.15.2",
         "webpack-merge": "^5.10.0"
     }
 }
