@@ -2,23 +2,23 @@
     "name": "webhost",
     "version": "1.0.0",
     "description": "A trivial host to serve static content and deal with web back end concerns",
-    "main": "server.ts",
-    "engines": {
-        "node": ">=18"
-    },
-    "scripts": {
-        "build": "rimraf dist && tsc",
-        "start": "npm run build && node dist/server.js"
-    },
     "author": "Curity AB",
     "license": "Apache-2.0",
+    "main": "server.ts",
+    "type": "module",
+    "engines": {
+        "node": ">=20"
+    },
+    "scripts": {
+        "build": "rm -rf dist && tsc",
+        "start": "npm run build && node dist/server.js"
+    },
     "dependencies": {
-        "express": "^4.18.2"
+        "express": "^4.19.2"
     },
     "devDependencies": {
-        "@types/express": "^4.17.17",
-        "@types/node": "^18.16.19",
-        "rimraf": "^3.0.2",
-        "typescript": "^5.1.5"
+        "@types/express": "^4.17.21",
+        "@types/node": "^18.19.33",
+        "typescript": "^5.4.5"
     }
 }
